backend/requirements.txt
PyYAML
flask
requests
uwsgi
beautifulsoup4
lxml
